Adjust values in graph test files for clarity
In the "util/test/test_graphs_folder" directory, the values connecting vertices 4 and 5 in the files graph1.txt, graph2.txt and graph3.txt were changed from 2 to -1. This change is intended to align with the updates in the graph testing process, specifically testing the created graphs and calculation of graph metrics. Now, the test files clearly reflect non-positive graph weights and make the behavior of related graph methods more predictable and consistent.