Add export functionality to instance parser
Added a new function 'export_properties' to the InstanceParserInterface class, allowing the properties of a graph to be exported to a file after its generation and anonymization. This will facilitate further data analysis steps by providing easy access to essential graph metrics such as number of vertices, edges, density, degree averages, and more. The main functionality of the module 'wikipedia_adminiship_election/code/main.py' was also adjusted in order to utilize the new export capabilities.